+++ src/main/java/com/invadermonky/bakedenchants/mixins/ContainerRepairMixin.java
@@ -1,6 +1,7 @@
 package com.invadermonky.bakedenchants.mixins;
 
-import com.invadermonky.bakedenchants.config.ConfigTags;
+import com.invadermonky.bakedenchants.handler.BakedEnchantmentHandler;
+import com.invadermonky.bakedenchants.handler.BakedEnchantmentRecipe;
 import com.llamalad7.mixinextras.sugar.Local;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.inventory.ContainerRepair;
@@ -13,7 +14,8 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 public class ContainerRepairMixin {
     @Redirect(method = "updateRepairOutput", at = @At(value = "INVOKE", target = "Lnet/minecraft/enchantment/Enchantment;isCompatibleWith(Lnet/minecraft/enchantment/Enchantment;)Z"))
     private boolean updateRepairOutputMixin(Enchantment applyEnchant, Enchantment itemEnchant, @Local(ordinal = 0) ItemStack stack) {
-        if (!applyEnchant.isCompatibleWith(itemEnchant) && ConfigTags.isBakedEnchant(stack, itemEnchant)) {
+        BakedEnchantmentRecipe recipe = BakedEnchantmentHandler.getBakedEnchantRecipe(stack);
+        if (recipe != null && !applyEnchant.isCompatibleWith(itemEnchant) && recipe.isBakedEnchant(stack, itemEnchant)) {
             return true;
         }
         return applyEnchant.isCompatibleWith(itemEnchant);