+++ src/main/java/com/invadermonky/bakedenchants/mixins/ItemStackMixin.java
@@ -1,8 +1,8 @@
 package com.invadermonky.bakedenchants.mixins;
 
-import com.invadermonky.bakedenchants.config.ConfigHandler;
-import com.invadermonky.bakedenchants.config.ConfigTags;
-import com.invadermonky.bakedenchants.util.EnchantmentHelperHelper;
+import com.invadermonky.bakedenchants.handler.BakedEnchantmentHandler;
+import com.invadermonky.bakedenchants.handler.BakedEnchantmentRecipe;
+import com.invadermonky.bakedenchants.util.EnchantmentHelperBE;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.EnchantmentHelper;
 import net.minecraft.init.Items;
@@ -31,14 +31,17 @@ public class ItemStackMixin {
 
     @Inject(method = "Lnet/minecraft/item/ItemStack;<init>(Lnet/minecraft/item/Item;IILnet/minecraft/nbt/NBTTagCompound;)V", at = @At("TAIL"))
     private void constructorMixin(Item itemIn, int amount, int meta, @Nullable NBTTagCompound capNBT, CallbackInfo ci) {
-        if (ConfigHandler.injectItemCreation && itemIn != null && itemIn != Items.AIR && ConfigTags.hasBakedEnchants(itemIn)) {
-            ConfigTags.addBakedEnchants(getItemStack());
+        if (itemIn != null && itemIn != Items.AIR) {
+            BakedEnchantmentRecipe recipe = BakedEnchantmentHandler.getBakedEnchantRecipe(getItemStack());
+            if (recipe != null) {
+                recipe.bakeEnchantments(getItemStack());
+            }
         }
     }
 
     @Inject(method = "isItemEnchanted", at = @At("HEAD"), cancellable = true)
     private void isItemEnchantedMixin(CallbackInfoReturnable<Boolean> ci) {
-        if (ConfigTags.canBeEnchanted(getItemStack())) {
+        if (!BakedEnchantmentHandler.canBeEnchanted(getItemStack())) {
             ci.setReturnValue(false);
         }
     }
@@ -46,21 +49,23 @@ public class ItemStackMixin {
     @Inject(method = "addEnchantment", at = @At("HEAD"), cancellable = true)
     private void addEnchantmentMixin(Enchantment enchToAdd, int level, CallbackInfo ci) {
         ItemStack stack = this.getItemStack();
-        int enchId = Enchantment.getEnchantmentID(enchToAdd);
-        int currLevel = EnchantmentHelper.getEnchantmentLevel(enchToAdd, stack);
-        if (currLevel > 0) {
-            if (currLevel < level) {
-                NBTTagList nbttaglist = stack.getTagCompound().getTagList("ench", 10);
-                for (NBTBase nbtBase : nbttaglist) {
-                    if (nbtBase instanceof NBTTagCompound && ((NBTTagCompound) nbtBase).getShort("id") == enchId) {
-                        ((NBTTagCompound) nbtBase).setShort("lvl", (byte) level);
+        if (BakedEnchantmentHandler.isBakedEnchantItem(stack)) {
+            int enchId = Enchantment.getEnchantmentID(enchToAdd);
+            int currLevel = EnchantmentHelper.getEnchantmentLevel(enchToAdd, stack);
+            if (currLevel > 0) {
+                if (currLevel < level) {
+                    NBTTagList nbttaglist = stack.getTagCompound().getTagList("ench", 10);
+                    for (NBTBase nbtBase : nbttaglist) {
+                        if (nbtBase instanceof NBTTagCompound && ((NBTTagCompound) nbtBase).getShort("id") == enchId) {
+                            ((NBTTagCompound) nbtBase).setShort("lvl", (byte) level);
+                        }
                     }
                 }
+                ci.cancel();
+            } else {
+                Map<Enchantment, Integer> currEnchants = EnchantmentHelper.getEnchantments(stack);
+                currEnchants.keySet().stream().filter(enchant -> !enchToAdd.isCompatibleWith(enchant)).forEach(enchant -> EnchantmentHelperBE.removeEnchantment(stack, enchant));
             }
-            ci.cancel();
-        } else {
-            Map<Enchantment, Integer> currEnchants = EnchantmentHelper.getEnchantments(stack);
-            currEnchants.keySet().stream().filter(enchant -> !enchToAdd.isCompatibleWith(enchant)).forEach(enchant -> EnchantmentHelperHelper.removeEnchantment(stack, enchant));
         }
     }
 }