+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 # Changelog
+## 1.12.2-2.0.0
+The 2.0.0 version of the mod featured a complete rebuild of much of the internal handling to support the improved item baking method.
+### Added
+- Added missing GroovyScript generated wiki documentation
+### Changed
+- Rebuilt mod with focus on ItemStack creation baking
+- Configuration item baking has been moved to Crafttweaker/GroovyScript integration
+### Removed
+- Removed Item Registry baking method
+
+---
+
 ## 1.12.2-1.1.0
 ### Added
 - Added mob drop baked enchantment injection (configurable)
@@ -7,6 +19,8 @@
 
 ### Fixed
 - Fixed JEI log error spam
+
+---
 
 ## 1.12.2-1.0.0
 - Initial Release